requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
attrs==25.3.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.3
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.2.2
faiss-cpu==1.10.0
fastapi==0.115.6
filetype==1.2.0
frozenlist==1.5.0
google-ai-generativelanguage==0.6.16
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.1
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
kaitaistruct==0.10
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.45
langchain-google-genai==2.1.0
langchain-openai==0.3.8
langchain-text-splitters==0.3.6
langgraph==0.3.11
langgraph-checkpoint==2.0.20
langgraph-prebuilt==0.1.3
langgraph-sdk==0.1.57
langsmith==0.3.15
marshmallow==3.26.1
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector-python==9.2.0
mysqlclient==2.2.7
numpy==2.2.3
openai==1.66.3
orjson==3.10.15
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydivert==2.1.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
selenium==4.30.0
selenium-wire==5.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.39
starlette==0.41.3
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.32.1
watchgod==0.8.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.18.3
zstandard==0.23.0
